pyproject.toml
[build-system]
requires = ["setuptools>=70.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-posty"
authors = [{name = "Simon Cam", email = "scam@wearecrew.com"}]
readme = "Readme.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version"]
dependencies = ["Django>=4.2"]
description = "A Django app to manage emails and templates"

[project.optional-dependencies]
extras = ["django_ace>=1.3.2", "django-flat-json-widget>=0.2"]

[project.urls]
Home = "https://github.com/wearecrew/django-posty"

[tool.setuptools.package-dir]
posty = "posty"

[tool.setuptools.package-data]
posty = ["templates/posty/*", "static/posty/*"]